refactor(inference): extract switch coordination
